# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ISP Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List all roles
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a new role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    get:
      operationId: getRole
      tags: [roles]
      summary: Get a role by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRole
      tags: [roles]
      summary: Update a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all zones
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneInput"
      responses:
        "201":
          description: Zone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
...
  /mikrotiks:
    get:
      operationId: listMikrotiks
      tags: [mikrotiks]
      summary: List all MikroTik routers with client stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: MikroTik list and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MikrotiksResponse"
    post:
      operationId: createMikrotik
      tags: [mikrotiks]
      summary: Add a MikroTik router
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MikrotikInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mikrotik"
  /mikrotiks/{id}:
    patch:
      operationId: updateMikrotik
      tags: [mikrotiks]
      summary: Update a MikroTik router
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MikrotikInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mikrotik"
        "404":
          description: Not found
    delete:
      operationId: deleteMikrotik
      tags: [mikrotiks]
      summary: Delete a MikroTik router
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Role:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        createdAt:
          type: string
          format: date-time
    RoleInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        permissionIds:
          type: array
          items:
            type: integer
    RoleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        permissionIds:
          type: array
          items:
            type: integer
    Permission:
      type: object
      required: [id, name, module, action]
      properties:
        id:
          type: integer
        name:
          type: string
        module:
          type: string
        action:
          type: string
        description:
          type: ["string", "null"]
...
    Zone:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ZoneInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    Package:
      type: object
      required: [id, name, price, createdAt, activeClients, inactiveClients, totalClients]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: string
        speed:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        activeClients:
          type: integer
        inactiveClients:
          type: integer
        totalClients:
          type: integer
    PackageInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: string
        speed:
          type: string
        description:
          type: string
...
    Mikrotik:
      type: object
      required: [id, name, publicIp, login, password, autoMkSync, autoSync, activeGraph, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        publicIp:
          type: string
        login:
          type: string
        password:
          type: string
        autoMkSync:
          type: boolean
        autoSync:
          type: boolean
        activeGraph:
          type: boolean
        webPort:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MikrotikInput:
      type: object
      required: [name, publicIp, login, password]
      properties:
        name:
          type: string
        publicIp:
          type: string
        login:
          type: string
        password:
          type: string
        autoMkSync:
          type: boolean
        autoSync:
          type: boolean
        activeGraph:
          type: boolean
        webPort:
          type: integer
        note:
          type: string
        status:
          type: string
    MikrotikStats:
      type: object
      required: [totalClients, activeClients, inactiveClients, onlineClients]
      properties:
        totalClients:
          type: integer
        activeClients:
          type: integer
        inactiveClients:
          type: integer
        onlineClients:
          type: integer
    MikrotiksResponse:
      type: object
      required: [mikrotiks, stats]
      properties:
        mikrotiks:
          type: array
          items:
            $ref: "#/components/schemas/Mikrotik"
        stats:
          $ref: "#/components/schemas/MikrotikStats"
...