# lib/api-zod/src/generated/types/mikrotik.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Mikrotik {
  id: number;
  name: string;
  publicIp: string;
  login: string;
  password: string;
  autoMkSync: boolean;
  autoSync: boolean;
  activeGraph: boolean;
  /** @nullable */
  webPort?: number | null;
  /** @nullable */
  note?: string | null;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}