# lib/api-zod/src/generated/types/mikrotiksResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Mikrotik } from './mikrotik';
import type { MikrotikStats } from './mikrotikStats';

export interface MikrotiksResponse {
  mikrotiks: Mikrotik[];
  stats: MikrotikStats;
}