# lib/api-zod/src/generated/types/zoneInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ZoneInput {
  name: string;
  description?: string;
}